+++ tests/src/sanity.rs
@@ -383,8 +383,17 @@ mod tests {
 
     #[cfg(target_os = "macos")]
     fn resolve_node_host() -> String {
-        // We assume it's Docker for Mac
-        "127.0.0.1".to_string()
+        if std::env::var("USE_MINIKUBE").is_ok() {
+            let output = std::process::Command::new("minikube")
+                .arg("ip")
+                .output()
+                .unwrap()
+                .stdout;
+            String::from_utf8_lossy(&output).to_string()
+        } else {
+            // We assume it's Docker for Mac
+            "127.0.0.1".to_string()
+        }
     }
 
     #[cfg(target_os = "linux")]